fix timestamp truncation

+++ firmware/ruler_lib/src/RulerButtons.h
@@ -33,7 +33,7 @@ private:
   const unsigned long c_debounceDelay = 75;  // the debounce time in ms
   int m_btn_pin[NUM_BTNS] = { INVALID_PIN, INVALID_PIN, INVALID_PIN, INVALID_PIN, INVALID_PIN };
   uint8_t m_last_btn_rd[NUM_BTNS] = { BTN_DEASSERT_LEVEL, BTN_DEASSERT_LEVEL, BTN_DEASSERT_LEVEL, BTN_DEASSERT_LEVEL, BTN_DEASSERT_LEVEL };  //raw read
-  uint8_t m_last_btn_debounce_time[NUM_BTNS] = { 0 };
+  uint32_t m_last_btn_debounce_time[NUM_BTNS] = { 0 };
   uint8_t m_debounce_btn_val[NUM_BTNS] = { 0 };  //this will be ASSERT=1, DEASSERT=0
 };
